+++ src/lib/components/App/App.test.tsx
@@ -1,13 +1,15 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import { act } from "react-dom/test-utils";
 
 import { App } from "./App";
+import { mockFetch } from '../../test-helpers/mock-fetch'
 
 describe("App", () => {
+    const mockFetchResponse = [{}];
+    window.fetch = mockFetch(mockFetchResponse);
     test("renders the App component", () =>
-        act(() => {
+        React.act(() => {
             render(<MemoryRouter><App /></MemoryRouter>);
         })
     )