+++ src/lib/components/SuperAdminOptions/GeneralConfig.test.tsx
@@ -1,13 +1,15 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
-import { act } from "react-dom/test-utils";
 
 import { GeneralConfig } from "./GeneralConfig";
+import { mockFetch } from '../../test-helpers/mock-fetch'
 
 describe("GeneralConfig", () => {
+    const mockFetchResponse = [{}];
+    window.fetch = mockFetch(mockFetchResponse);
     test("renders the GeneralConfig component", () =>
-        act(() => {
+        React.act(() => {
             render(<MemoryRouter><GeneralConfig /></MemoryRouter>);
         })
     )