+++ src/lib/test-helpers/mock-fetch.ts
@@ -1,0 +1,13 @@
+export function mockFetch(data: any, headers: any = null) {
+    if (!headers) {
+        headers = {'Content-Type': 'application/json'};
+    }
+    return jest.fn().mockImplementation(() =>
+      Promise.resolve({
+        ok: true,
+        json: () => data,
+        headers: new Headers(headers),
+        text: () => Promise.resolve(JSON.stringify(data)),
+      }),
+    );
+}