+++ CHANGELOG.md
@@ -16,14 +16,16 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
-## Unreleased
-## 1.0.21 (2024-06-06)
+## 1.0.21 (2024-07-19)
 
 ### Fixes
 Fix: error "Failed to execute 'atob' on 'Window': The string to be decoded contains characters outside of the Latin1" [GS-95].
+Fix: "ReferenceError: Response is not defined" error in npm test.
+Fix: "Warning: ReactDOMTestUtils.act is deprecated in favor of React.act" in npm test.
+Fix: "TypeError: (0 , _authenticationService.getUserData) is not a function" error in npm test.
 
 
-## 1.0.20 (2024-06-06)
+## 1.0.20 (2024-05-17)
 ---
 
 ### New